+++ test_web/index.html
@@ -27,7 +27,6 @@
                 <label class="input-group-text" for="language">Language</label>
             </div>
         </div>
-        <button onclick="fetch('/json')" class="btn btn-outline-success my-2 my-sm-0" type="button" id="clear-search">JSON plugins</button>
     </nav>
     <div id="test-container">
         <div class="test-wrapper" id="popularNovels">
